Use correct variable type to avoid truncation on 64-bit.
git-svn-id: https://svn.code.sf.net/p/quark/code@8555 5419a3ea-6bc3-475c-b79b-167d7c3fbd5f

+++ source/trunk/prog/QkTreeView.pas
@@ -1390,7 +1390,7 @@ end;
 procedure TMyTreeView.WMKillFocus(var Message: TMessage);
 begin
  HasFocus:=False;
- if (EditInfo=Nil) or (DWORD(Message.wParam)<>EditInfo^.InPlaceEdit.Handle) then
+ if (EditInfo=Nil) or (Message.wParam<>EditInfo^.InPlaceEdit.Handle) then
   begin
    CancelMouseClicking(True);
    Invalidate;
